requirements.txt
pyasn1~=0.4.8
pyOpenSSL~=19.1.0
Twisted~=20.3.0
# install twisted requests since they changed how that all works
treq~=20.3.0
pyparsing~=2.4.7
python-dateutil~=2.8.1
service_identity~=18.1.0
lockfile~=0.12.2

falcon~=2.0.0
txredisapi~=1.4.7
tabulate~=0.8.7

txAMQP @ git+https://github.com/jookies/txamqp.git@master
smpp.pdu @ git+https://github.com/jookies/smpp.pdu.git@master
smpp.twisted @ git+https://github.com/jookies/smpp.twisted.git@master
python-messaging @ git+https://github.com/jookies/python-messaging.git@master


# Added in 0.9rc16:
celery>=4.0.0
redis~=3.4.1
requests~=2.23.0

# For REST API
python-mimeparse~=1.6.0
